ready

+++ src/main/java/com/kateVoronina/demo/Constants.java
@@ -4,5 +4,5 @@ public class Constants {
 
     public static final String API_SECRET_KEY = "expensetrackerapikey";
 
-    public static final long TOKEN_VALIDITY = 2 * 60 * 60 * 1000;
+    public static final int TOKEN_VALIDITY = 2 * 60 * 60 * 1000;
 }
